renamed the project to vertx-auth-properties based on Paulo's  comments
