+++ pom.xml
@@ -30,7 +30,7 @@
     <module>vertx-auth-htdigest</module>
     <module>vertx-auth-htpasswd</module>
     <module>vertx-auth-webauthn</module>
-    <module>vertx-auth-file</module>
+    <module>vertx-auth-properties</module>
   </modules>
 
   <parent>
