--- vertx-auth-file/README.adoc
@@ -1,3 +1,0 @@
-# Vert.x Auth File
-
-This module contains contains a Vert.x auth implementation where users, roles and permisions are read from a text file
