+++ vertx-auth-properties/README.adoc
@@ -1,0 +1,3 @@
+# Vert.x Auth File
+
+This module contains contains a Vert.x auth implementation where users, roles and permissions are read from a property file
