+++ vertx-auth-properties/src/main/asciidoc/index.adoc
