+++ vertx-auth-properties/src/test/resources/test-auth.properties
